requirements.txt
fastapi
uvicorn
transformers
torch
tensorflow
pytesseract
opencv-python
google-cloud-vision
Pillow
